bdd: derive sibling feature prefixes from tags, not file names
d154dfc taught the clean/verify scans to skip resources owned by a
sibling feature whose tag extends this one's, but derived those sibling
prefixes from .feature file *stems*. Resources are named from each
feature's *tag*, and seven features declare a tag that differs from
their stem (isis_tilfa_srv6.feature is @tilfa_srv6, isis-redist.feature
is @isis_redist, ...). For any such sibling the stem-derived prefix
matches none of its real resource names, leaving it exposed to exactly
the clobbering d154dfc removed - silently, and only under concurrency.

Read the tag out of each sibling's file text instead, with the same
first-non-special-tag selection the cucumber before-hook applies to the
parsed feature. The parser lives in bdd::feature_tag so the harness and
the tag-uniqueness test share one definition.

This supersedes e8c03f5c, which kept the stem-based lookup and instead
grew tag_guard into a build-failing precondition check that every
tag-prefix pair's file names line up with its tags. Deriving from tags
makes that alignment irrelevant by construction, so the guard goes away
rather than being maintained: tag_guard is deleted, and its surviving
invariant - no two features may share a tag - moves to feature_tag's
tests. Behavior today is unchanged: the three real prefix pairs
(bgp_adv_interval_zero/_v6, /_vrf, bgp_evpn_vpws/_multihoming) all have
aligned file names, so old and new code compute identical sibling sets;
what changes is that a future misaligned pair works instead of failing
the build with a rename demand.
