+++ bdd/src/lib.rs
@@ -1,4 +1,3 @@
+pub mod feature_tag;
 pub mod netns;
 pub mod toolchain;
-
-mod tag_guard;
